Add basic support for anonymous shared memory in pseudo_mm. I have to say that maybe serverless application in our test won't use anonymous shared memory, but still have to support this feature for completeness.
For now, anonymous shared memory is implemented by a special file
in a special tmpfs inside kernel. So we can think it is a file-backed
mapping in a special tmpfs.

If we want to support anonymous shared memory in pseudo_mm, we need:
1. different attachment of pseudo_mm will be independent (i.e. they
  should not notice others' modification).
2. After an attachment, it should behave like a real shared memory.
  (e.g. after fork() parent and child should notice the modification).

In short, we need keep its content in COW manner in pseudo_mm level while
also provide the shared semantics after attachment.

So I decided to create a new special shmem file each time when attach pseudo_mm,
while let the page table entry (wrprotect) point to same page while attaching.
So at this point, the vm_file is not corresponding to the physical
pages (not-matching problem).

After a write request, it will cause page fault. We distinguish the
special vma in PF handler by adding a new field in vma called
`unsigned long pseudo_mm_flag`. Its lower 32-bit is pseudo_mm id and
higher 32-bit are bit flags.

Then we allocate a new page in our own shmem file and copy from the
previous page (already pointed by page table). From this point, the
vm_file matches with the physical page.

For the not-matching problem we add a refcount to prevent deleting the
shmem file while some process still point to it. (It can also be
solved by never deleting pseudo_mm but that is too brute force.)

And for the stability, I think it is better to pin the pages in filled
vma (i.e. the area that has been filled page's content in pseudo_mm) in
memory, to prevent swapping to create more subtle cases needed to
be handled in PF handler.