Add support for dax device as memory backend for pseudo_mm.
Now we support dax device (tested on Optane pmem). The physical pages
will be allocated on the dax device and managed by pseudo_mm module.

But for now it is only support anonymous private vma and I do not
decided to support anonymous shared vma for now since it is rare in
our serverless application.

The interface will change in next patch since current interface is only
useable for single host (not suitable for multi-host environment).