Add debug info to trace_printk() and couple of counters in mm.
To well understand the performance of pseudo_mm (with dax and rdma), I
need to get the latency of page fault. Since page fault is a highly
frequent and low-latency event, I choose the trace_printk().
Furthermore, I add couple of counters in mm to record the page fault
related to process that has attached the pseudo_mm. This counters can be
fetched from userspace using ioctl on /dev/pseudo_mm.